After_video/requirements.txt
whisperx
openai
python-dotenv
watchdog
torch
